Rename and move circuit module
