+++ app/build.gradle.kts
@@ -148,8 +148,8 @@ dependencies {
     implementation(projects.features.media)
     implementation(projects.features.pdf)
     implementation(projects.features.settings)
-    implementation(projects.libraries.circuitHelpers.impl)
     implementation(projects.libraries.foundation.coroutines)
+    implementation(projects.services.circuit.impl)
     implementation(projects.services.lessons.impl)
     implementation(projects.services.media.impl)
     implementation(projects.services.storage.impl)
