+++ features/account/build.gradle.kts
@@ -47,7 +47,7 @@ dependencies {
     implementation(projects.common.lessonsData)
     implementation(projects.common.models)
     implementation(projects.common.translations)
-    implementation(projects.libraries.circuitHelpers.api)
+    implementation(projects.libraries.circuit.api)
     implementation(projects.libraries.foundation.coroutines)
 
     implementation(libs.google.material)
