+++ libraries/circuit/api/src/main/kotlin/ss/libraries/circuit/navigation/CustomTabsIntentScreen.kt
@@ -1,4 +1,4 @@
-package ss.circuit.helpers.navigation
+package ss.libraries.circuit.navigation
 
 import com.slack.circuit.runtime.screen.Screen
 import kotlinx.parcelize.Parcelize
