+++ libraries/circuit/api/src/main/kotlin/ss/libraries/circuit/overlay/AlertDialogOverlay.kt
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package ss.circuit.helpers.overlay
+package ss.libraries.circuit.overlay
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.ExperimentalMaterial3Api
