+++ settings.gradle.kts
@@ -64,8 +64,7 @@ include(
     ":features:pdf",
     ":features:reader",
     ":features:settings",
-    ":libraries:circuit-helpers:api",
-    ":libraries:circuit-helpers:impl",
+    ":libraries:circuit:api",
     ":libraries:foundation:android",
     ":libraries:foundation:coroutines",
     ":libraries:foundation:coroutines:test",
@@ -82,6 +81,7 @@ include(
     ":libraries:test_utils",
     ":libraries:ui:placeholder",
     ":libraries:workers:api",
+    ":services:circuit:impl",
     ":services:lessons:impl",
     ":services:media:impl",
     ":services:storage:impl",
